More AR quoting changes

+++ activerecord/lib/active_record/associations/has_and_belongs_to_many_association.rb
@@ -70,7 +70,7 @@ module ActiveRecord
             records.each { |record| @owner.connection.delete(interpolate_sql(sql, record)) }
           else
             ids = quoted_record_ids(records)
-            sql = "DELETE FROM #{@owner.connection.quote_table_name @reflection.options[:join_table]} WHERE #{@reflection.primary_key_name} = #{owner_quoted_id} AND #{@reflection.association_foreign_key} IN (#{ids})"
+            sql = "DELETE FROM #{@owner.connection.quote_table_name @reflection.options[:join_table]} WHERE #{@owner.connection.quote_column_name @reflection.primary_key_name} = #{owner_quoted_id} AND #{@owner.connection.quote_column_name @reflection.association_foreign_key} IN (#{ids})"
             @owner.connection.delete(sql)
           end
         end
@@ -79,11 +79,11 @@ module ActiveRecord
           if @reflection.options[:finder_sql]
             @finder_sql = interpolate_sql(@reflection.options[:finder_sql])
           else
-            @finder_sql = "#{@owner.connection.quote_table_name @reflection.options[:join_table]}.#{@reflection.primary_key_name} = #{owner_quoted_id} "
+            @finder_sql = "#{@owner.connection.quote_table_name @reflection.options[:join_table]}.#{@owner.connection.quote_column_name @reflection.primary_key_name} = #{owner_quoted_id} "
             @finder_sql << " AND (#{conditions})" if conditions
           end
 
-          @join_sql = "INNER JOIN #{@owner.connection.quote_table_name @reflection.options[:join_table]} ON #{@reflection.quoted_table_name}.#{@reflection.klass.primary_key} = #{@owner.connection.quote_table_name @reflection.options[:join_table]}.#{@reflection.association_foreign_key}"
+          @join_sql = "INNER JOIN #{@owner.connection.quote_table_name @reflection.options[:join_table]} ON #{@reflection.quoted_table_name}.#{@owner.connection.quote_column_name @reflection.klass.primary_key} = #{@owner.connection.quote_table_name @reflection.options[:join_table]}.#{@owner.connection.quote_column_name @reflection.association_foreign_key}"
 
           if @reflection.options[:counter_sql]
             @counter_sql = interpolate_sql(@reflection.options[:counter_sql])
